requirements.txt
gunicorn
Flask==3.1.3
pyaml==26.2.1
pycountry==26.2.16
requests==2.32.5
deep-translator==1.11.4